Move reports code to reports dedicated file
